Data Entry  (#293)
* feat: Add DTOs and services for RenterMast, RoomWiseMinusData, and RoomWiseSubmissionDetails

- Introduced RenterMastDto and UpdateRenterMastDto for managing renter mast data.
- Created CreateRoomWiseMinusDataDto, RoomWiseMinusDataDto, and UpdateRoomWiseMinusDataDto for room-wise minus data handling.
- Added CreateRoomWiseSubmissionDetailsDto, RoomWiseSubmissionDetailsDto, and UpdateRoomWiseSubmissionDetailsDto for submission details management.
- Implemented IDataEntryService, IRenterDetailService, IRenterMastService, and IRoomWiseSubmissionDetailsService interfaces for CRUD operations.
- Developed DataEntryService to manage property details and associated entities.
- Created RenterDetailService, RenterMastService, and RoomWiseSubmissionDetailsService for handling respective entity operations.
- Added RenterMastEntity, RenterDetailEntity, RoomWiseMinusDataEntity, and RoomWiseSubmissionDetailsEntity to the core entities.
- Updated PropertyAssessmentEntity and PropertyDetailsEntity to include relationships with new entities.

* Refactor RoomWiseSubmissionDetails and related services

- Updated RoomWiseSubmissionDetailsDto and UpdateRoomWiseSubmissionDetailsDto to include non-nullable properties for OuterYesNo and MinusYesNo.
- Modified IRenterDetailService and IRenterMastService interfaces to use Create and Update DTOs.
- Refactored DataEntryService to streamline nested list handling for creation and updates.
- Implemented new mapping profiles for RenterDetail, RenterMast, RoomWiseMinusData, and RoomWiseSubmissionDetails.
- Adjusted entity definitions to remove unnecessary properties and ensure proper foreign key relationships.
- Enhanced the overall structure and clarity of the mapping configurations for better maintainability.

* feat: Add DataEntryController and related services for property details management

* change in Entity and ApplicationContext

* test

test

* Updates the changes to resolved the login Issue

Updates the changes to resolved the login Issue

* Resolved Id Issue and DbContextIssue

Resolved Id Issue and DbContextIssue

* Add changes to resolve casting error

Add changes to resolve casting error

* Updates changes related for filter by PropertyId

Updates changes related for filter by PropertyId

* Updates Related to the RoomWiseMinusData comes nullable

Resolved the RoomwiseMinusData insert issue

* updated related the roomWiseminusSubmission

updated related the roomWiseminusSubmission

* remove changes

remove changes

* changes related to update the created date

changes related to update the created date

* Added Seperate Crud

Added Seperate Crud  for RenterDetails,RenterMast,RoomWiseSubmission,RoomWiseMinus also added service for RoomWiseMinus

* Added hardDeletable fo pdn

Added hardDeletable fo pdn

* Added Changes for To Validate the MarkForDeletion

Added Changes for To Validate the MarkForDeletion

* resolved conflicts

resolved conflicts

* Update UserRepositories

Update UserRepositories

* added test cases for the DataEntrys

added test cases for the DataEntrys

* Resolved the Suggetions

Resolved the Suggetions

* updates the changes according the suggetions

updates the changes according the suggetions

* resolved copilet suggetions

resolved copilet suggetions

* Resolved the Copilet Suggetions

Resolved the Copilet Suggetions

* Resolved the copilet suggetions

Resolved the copilet suggetions

* Resolved The suggetions and build Errors

Resolved The suggetions and build Errors

* Resolved the copilet suggetions

Resolved the copilet suggetions

* Resolved Copilet suggetions

Resolved Copilet suggetions

* Resolved Suggetions

Resolved Suggetions

* Resolved copilet suggetions

Resolved copilet suggetions

* Resolved Suggetions

Resolved Suggetions

* Resolved Suggetions

Resolved Suggetions

* Updates the changes

Updates the changes

* Resolved Test Errors

Resolved Test Errors

* Resolved the suggetions

Resolved the suggetions

* updates the changes

updates the changes

* Added Latest Changes

Added Latest Changes

* Resolved All suggetions

Resolved All suggetions

* Added changes for MarkForDeletion

Added changes for MarkForDeletion

* Resolved conflicts

Resolved conflicts

* resolved conflicts

resolved conflicts

* Resolved conflicts Issues

Resolved conflicts Issues

* Resolved Conflicts Errors

Resolved Conflicts Errors

* Resolved the Build Error

Resolved the Build Error

* Resolved Build Errors

Resolved Build Errors

* Resolved Test Cases Errors

Resolved Test Cases Errors

* Resolved Test Errors

Resolved Test Errors

* Resolved The Suggetions

Resolved The Suggetions

* Resolved the Test errors

Resolved the Test errors

* Resolved the suggetions

Resolved the suggetions

* change the decimal to double

change the decimal to double

* Resolved the dataType Errors for decimal to double

* Resolved Suggetions

Resolved Suggetions

---------

+++ src/Core/NtisPlatform.Core/Entities/RoomWiseMinusDataEntity.cs
@@ -12,17 +12,17 @@ public class RoomWiseMinusDataEntity : BaseEntity, IHardDeletable
 {
     public int RoomWiseSubmissionId { get; set; }
 
-    public decimal? LengthMtr { get; set; }
+    public double? LengthMtr { get; set; }
 
-    public decimal? WidthMtr { get; set; }
+    public double? WidthMtr { get; set; }
 
-    public decimal? AreaSqMtr { get; set; }
+    public double? AreaSqMtr { get; set; }
 
-    public decimal? HeightMtr { get; set; }
+    public double? HeightMtr { get; set; }
 
-    public decimal? Base1Mtr { get; set; }
+    public double? Base1Mtr { get; set; }
 
-    public decimal? Base2Mtr { get; set; }
+    public double? Base2Mtr { get; set; }
 
     public string? Shape { get; set; }
